backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.5
python-socketio==5.8.0
python-engineio==4.8.0
eventlet==0.33.3
